zadanie: rozbudowa testów kalkulatora

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Application.java
@@ -4,10 +4,10 @@ public class Application {
     public static void main(String[] args) {
 
         Calculator calculator = new Calculator();
-        int a = 5;
-        int b = 8;
+        double a = 5;
+        double b = 8;
 
-        int sumResult = calculator.sum(a, b);
+        double sumResult = calculator.sum(a, b);
 
         boolean correct = ResultChecker.assertEquals(13, sumResult);
 
@@ -17,7 +17,7 @@ public class Application {
             System.out.println("Metoda sum nie działa poprawnie dla liczb " + a + " i " + b);
         }
 
-        int substractResult = calculator.substract(a,b);
+        double substractResult = calculator.substract(a,b);
 
         boolean correctTwo = ResultChecker.assertEquals(-3, substractResult);
 
@@ -27,7 +27,7 @@ public class Application {
             System.out.println("Metoda nie wykonala poprawnie odejmowania libcz " + a + " i " + b);
         }
 
-        int squareResult = calculator.square(a);
+        double squareResult = calculator.square(a);
 
         boolean correctThree = ResultChecker.assertEquals(25, squareResult);
